+++ src/entity/user.ts
@@ -8,3 +8,13 @@ export interface User {
   createdAt: Date;
   updatedAt: Date;
 }
+
+export interface SignUpFormInput
+  extends Pick<User, "name" | "email" | "companyName"> {
+  password: string;
+  passwordCheck: string;
+}
+
+export interface LoginFormInput extends Pick<User, "email"> {
+  password: string;
+}
